+++ src/Entities/Shared/AdjustmentTimePeriod.php
@@ -9,7 +9,7 @@ declare(strict_types=1);
  * |-------------------------------------------------------------|.
  */
 
-namespace Paddle\SDK\Entities\Adjustment;
+namespace Paddle\SDK\Entities\Shared;
 
 use Paddle\SDK\Entities\DateTime;
