+++ tests/Functional/Resources/Adjustments/AdjustmentsClientTest.php
@@ -7,12 +7,12 @@ namespace Paddle\SDK\Tests\Functional\Resources\Adjustments;
 use GuzzleHttp\Psr7\Response;
 use Http\Mock\Client as MockClient;
 use Paddle\SDK\Client;
-use Paddle\SDK\Entities\Adjustment\AdjustmentItem;
-use Paddle\SDK\Entities\Adjustment\AdjustmentType;
 use Paddle\SDK\Entities\Shared\Action;
 use Paddle\SDK\Entities\Shared\AdjustmentStatus;
+use Paddle\SDK\Entities\Shared\AdjustmentType;
 use Paddle\SDK\Environment;
 use Paddle\SDK\Options;
+use Paddle\SDK\Resources\Adjustments\Operations\Create\AdjustmentItem;
 use Paddle\SDK\Resources\Adjustments\Operations\CreateAdjustment;
 use Paddle\SDK\Resources\Adjustments\Operations\ListAdjustments;
 use Paddle\SDK\Resources\Shared\Operations\List\Pager;
